Changed int to Double

+++ Week 2 - Arrays/Score/scores.c
@@ -4,9 +4,9 @@
 int main(void)
 {
     // Taking User Input of Scores
-    int score1 = get_int("Give me a Score: ");
-    int score2 = get_int("Give me a Score: ");
-    int score3 = get_int("Give me a Score: ");
+    double score1 = get_int("Give me a Score: ");
+    double score2 = get_int("Give me a Score: ");
+    double score3 = get_int("Give me a Score: ");
 
     // Calculating and Printing the Average
     printf("Average Score: %f\n", (score1 + score2 + score3) /3);
